+++ shuup/importer/importing/importing.py
@@ -21,6 +21,7 @@ from django.utils.text import force_text
 from django.utils.translation import ugettext_lazy as _
 from enumfields import EnumIntegerField
 
+from shuup.admin.utils.permissions import has_permission
 from shuup.importer._mapper import RelatedMapper
 from shuup.importer.exceptions import ImporterError
 from shuup.importer.importing.meta import ImportMetaBase
@@ -55,9 +56,10 @@ class DataImporter(object):
 
     model = None
 
-    def __init__(self, data, shop, language):
+    def __init__(self, data, shop, language, supplier=None):
         self.shop = shop
         self.data = data
+        self.supplier = supplier
         self.data_keys = data[0].keys()
         self.language = language
 
@@ -68,7 +70,7 @@ class DataImporter(object):
         self.field_defaults = self._meta.get_import_defaults()
 
     @classmethod
-    def transform_file(cls, mode, filename, data=None):
+    def transform_file(cls, mode, filename, data=None, supplier=None):
         """
         That method will be called if `cls.custom_file_transformer` is True
         """
@@ -145,7 +147,6 @@ class DataImporter(object):
 
             elif not mapped_value and not self._meta.has_post_save_handler(field_name):
                 self.unmatched_fields.add(field_name)
-
         self.data_map = data_map
         return data_map
 
@@ -253,6 +254,10 @@ class DataImporter(object):
         row_lower = {key.lower(): val for key, val in row.items()}
         if row_lower.get("ignore"):
             return
+
+        if self.supplier:
+            if "Supplier" in row.keys():
+                row["Supplier"] = self.supplier
 
         obj, new = self._resolve_obj(row)
         if not obj: