+++ CHANGELOG.md
@@ -24,6 +24,7 @@ List all changes after the last release here (newer on top). Each change on a se
 - Core: add shops and supplier to ServiceProvider and Service models
 - Front: add feature for checkout phases to spawn extra phases
 - Add custom get_ip method and use it everywhere
+- Importer: add permissions for all the diffrent types of importers
 
 ### Fixed